requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
amqp==5.3.1
    # via kombu
asgiref==3.9.1
    # via
    #   channels
    #   channels-redis
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
celery==5.5.3
    # via -r requirements.in
certifi==2025.8.3
    # via requests
cffi==2.0.0
    # via cryptography
channels==4.3.1
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.3.0
    # via -r requirements.in
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
cryptography==45.0.7
    # via jwcrypto
django==5.2.6
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   django-extensions
    #   django-oauth-toolkit
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-spectacular
django-cors-headers==4.8.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-oauth-toolkit==3.0.1
    # via -r requirements.in
django-redis==6.0.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-nested-routers==0.95.0
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
jwcrypto==1.5.6
    # via django-oauth-toolkit
kombu==5.5.4
    # via celery
msgpack==1.1.1
    # via channels-redis
numpy==2.3.3
    # via pgvector
oauthlib==3.3.1
    # via django-oauth-toolkit
packaging==25.0
    # via kombu
pgvector==0.4.1
    # via -r requirements.in
pillow==11.3.0
    # via -r requirements.in
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.23
    # via cffi
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   celery
python-decouple==3.8
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
pytz==2025.2
    # via -r requirements.in
pyyaml==6.0.2
    # via drf-spectacular
redis==6.4.0
    # via
    #   -r requirements.in
    #   channels-redis
    #   django-redis
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via django-oauth-toolkit
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via
    #   jwcrypto
    #   referencing
tzdata==2025.2
    # via kombu
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
